run clippy
